Clean the ChannelWatcher, make it actually useful
Now it monitors the channel and on any message it constructs the
StorageMessage struct that it sends to the TextChannelObserver for
storage.

The more I work on this class the more I think it is not really needed.
I think I'll scrap it in the near future and move the processing to the
observer itself.