Fixed a little bug that make Spacebar unable to be invoked as sms handler application by other programs
Added the mimetype, so that Spacebar is recognized by the system as a valid sms handler application.
Changed the Exec line to `spacebar %u`. The missing "%u" cause the error `klauncher said: Unknown protocol 'sms'` when opened by a "sms link" on Plasma Mobile. That error could be triggered with`xdg-open sms:"phone number"` or by the `plasma-phonebook` application by clicking on `Write SMS on "phone number"` from any contact.