Properly order chat list homepage
Currently, the `lastContacted` field for numbers isn't actually the
right date: instead, it is some random date within a month of 1970.
This causes the chat list homepage to display chats in a random order.

The timestamp is stored into the database as a qint64, so it should be
read out the same way. This changes the way that the number is taken
from the database to eliminate the issue.