Prevent last message text from overflowing
Sometimes the last chat message would overflow the available space, and
go past the end of the line due to the default wrapMode allowing the
content width to exceed the set width when there are insufficient
newlines. This changes the wrapping strategy to aggresively wrap the
text to fit within the available space.

The effect of using WrapAnywhere instead of Wrap is that some
characters of a cut-off word will be displayed instead of wrapping
earlier.