Fix failed assert when messaging non-contact
In src/messagemodel.cpp, sendMessage gets the most recent message:
    result = co_await sendMessageInternal(m_phoneNumberList.first(), text);
This fails if m_phoneNumberList is empty (and thus has no first
element). The lack of phone numbers is ultimately caused in
src/contents/ui/NewConversationPage.qml, which doesn't include the
search field when creating a conversation. This fixes it to include the
search as a phone number if no contacts are selected.

Closes #35.