Improve how the members of a conversation are displayed
Show how many more people are in conversation if all members cannot be displayed in the chats list and also on the page header of the current conversation.

Added page to display all the members in current conversation.

Use the contacts list instead of displaying chips at the top of a conversation for listing, adding, and/or removing people.

![spacebar-members-blur](/uploads/7b9691a2ccfb3472c86b11023c65a0c7/spacebar-members-blur.png)

Fixes [#54](https://invent.kde.org/plasma-mobile/spacebar/-/issues/54)