Rework flatpak for MM
This is not tested on a real device, and it's very possible it doesn't
completely work.
The daemon likely doesn't work, I'm not sure how something like that
should be done with flatpak.

It would be cool if we could publish that to flathub in the end, and
this goes a large part of that way.