Rework and improve new chat page UX
I found the existing "new chat" page UX to be very hard to figure out,
so I reworked it to distinguish between "single contact conversations"
and "groups". Initially, when the conversation page opens, selecting a
contact will immediately create a chat with them. Selecting the "new
group" option will put the user into selection mode, with a button to
create the group.