Remove all SQL queries from client, move to DBus only
Fixes #66

We seem to be getting conflicts due to the client and daemon both accessing the same sqlite file at the same time. Move to have all client requests be done over DBus instead of direct SQL accesses. This DBus API is an exact translation of the db calls at the moment. I had to remove the auto-generated dbus file (now it's manually specified) because it wasn't properly translating some interface functions. 

The current major caveat is that the requests are blocking again on the UI, it can be made async again in the future but I whipped this up quite quickly to have this done....